Many changes. Details after the break.
A big thing is that I've moved a lot of stuff into individual files, and
started moving to a more class-based approach. summarize() probably does
not work at the moment, however it did get as far as working for the
yearly and quarterly data. It is likely that it will go through even
larger changes with the addition of classes.

There is also the getDateRanges() function in util.js, which works very
nicely for getting ranges for every quarter, month, and week that have
passed so far in the current year. Also, naturally, the year itself.